+++ graphql/users.graphql
@@ -1,0 +1,7 @@
+query users {
+  users {
+    id
+    last_seen
+    name
+  }
+}
