+++ hasura/metadata/databases/default/tables/public_profiles.yaml
@@ -1,0 +1,3 @@
+table:
+  name: profiles
+  schema: public
