+++ hasura/metadata/databases/default/tables/tables.yaml
@@ -1,2 +1,2 @@
-- "!include public_profile.yaml"
+- "!include public_profiles.yaml"
 - "!include public_users.yaml"
